requirements.txt

charset-normalizer  3.4.2
dotenv              0.9.9
numpy               2.2.6
pandas              2.2.3
python-dotenv       1.1.0
python-telegram-bot 22.1                
requests            2.32.3

